pyproject.toml
[project]
name = "temporal-otel"
version = "0.1.0"
description = "Sample for working with Temporal OpenTelemetry integration."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp==3.12.14",
    "opentelemetry-api==1.35.0",
    "opentelemetry-instrumentation-aiohttp-client==0.56b0",
    "opentelemetry-sdk==1.35.0",
    "temporalio==1.14.1",
]

[dependency-groups]
dev = [
    "poethepoet>=0.36.0",
    "ruff>=0.12.5",
]

[tool.poe.tasks.worker]
help = "Run the worker."
cmd = "python -m python.worker"
env = { PYTHONPATH = "." }

[tool.poe.tasks.workflow]
help = "Run the workflow."
cmd = "python -m python.workflow"
env = { PYTHONPATH = "." }
